requirements.txt
# Core testing framework
pytest>=7.4.0
playwright>=1.40.0

# Pytest plugins
pytest-playwright>=0.4.3
pytest-xdist>=3.3.1

# Test data generation
faker>=19.0.0

# Reporting and allure
allure-pytest>=2.13.2

# Additional utilities
python-dotenv>=1.0.0
